+++ Algorithm/KindOperationType.go
@@ -1,6 +1,7 @@
 package Algorithm
 
-func kindOperationType(answer, productValue string) (diff int) {
+//kindOperationType calculates the influence of an answer given to a replace/support lecture question.
+func kindOperationType(answer, productValue string) (diff int8) {
 
 	switch ans := answer; ans {
 	case `support4lecture`:
